requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
carla==0.9.15
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
coverage==7.6.1
cryptography==42.0.4
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
Django==4.2.16
djangorestframework==3.15.2
dnspython==2.6.1
email-validator==2.1.1
exceptiongroup==1.2.1
executing==2.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
filelock==3.16.1
flask==3.0.3
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.35.1
importlib==1.0.4
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
imageio==2.35.1
imageio-ffmpeg==0.5.1
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.13
kiwisolver==1.4.7
lapx==0.5.10.post1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.1
numpy==1.24.4
open3d==0.18.0
opencv-python==4.10.0.84
orjson==3.10.4
packaging==24.1
pandas==2.0.3
parso==0.8.4
pem==23.1.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prompt-toolkit==3.0.48
psutil==6.0.0
psycopg2-binary==2.9.9
pure-eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.7.3
pydantic-core==2.18.4
pydantic-extra-types==2.8.0
pydantic-settings==2.3.1
pygame==2.6.0
pygments==2.18.0
pyotp==2.9.0
pyparsing==3.1.4
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.20.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tomli==2.0.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.3
typing-extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
ultralytics==8.2.103
ultralytics-thop==2.0.8
urllib3==2.2.3
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
werkzeug==3.0.4
widgetsnbextension==4.0.13
yarl==1.13.1
zipp==3.20.2